requirements/developer.txt
# Install the base set of requirements
-r ./base.txt

# Additional requirements useful for developers of the webapp

# Django debugging
django-debug-toolbar

# Testing
tox
selenium
browserstack-local

# docker compatible debugger
pudb
